pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pydantic-extra-types (>=2.10.3,<3.0.0)",
    "phonenumbers (>=9.0.2,<10.0.0)",
    "boto3 (>=1.37.27,<2.0.0)"
]


[project.scripts]
main_app = "server.main:app"


[tool.poetry]
packages = [{include = "server", from = "src"}]


[tool.pytest.ini_options]
pythonpath = "src"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
